refactor(calendar): extraire useCalendarEvents d'UniversalCalendar (#28bis)
Ramène le <script> d'UniversalCalendar.vue de 443 à 242 lignes (<300,
PRODUCTION_STANDARDS §1.1) en déplaçant l'état (events/matieres/classes/
enseignants + loading/refreshing) et le chargement (séances + évaluations
selon le rôle, options de filtres dérivées, refresh KLASSCI bypass cache)
dans le composable composables/useCalendarEvents.js.

Comportement identique : mêmes appels lmsService/evaluationService, mêmes
couleurs/urgence/bornes (utils/calendar, testés). refreshEvents() délègue
désormais à refreshData() qui force le bypass cache des 2 sources (l'ancien
loadSeances(true) jetait son résultat avant un loadEvents() non forcé).

Vérifié : vite build vert, tests/unit/CalendarFilters.test.js vert (5/5).
